requirements.txt
# ACP Polymarket Trading Agent Requirements

# Core dependencies
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
asyncio==3.4.3
websockets==11.0.3
httpx==0.25.0
aiohttp==3.8.6
sqlalchemy==2.0.22
alembic==1.12.0
loguru==0.7.2

# ACP SDK
acp-sdk==0.1.0  # Replace with actual ACP SDK version when available

# Blockchain & Wallet
web3==6.10.0
eth-account==0.9.0
eth-typing==3.4.0
eth-utils==2.2.0

# Trading & Analytics
numpy==1.26.0
pandas==2.1.1
ta-lib==0.4.28  # Technical analysis library
scikit-learn==1.3.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Deployment
docker==6.1.3
gunicorn==21.2.0

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
